Build dependencies with Docker
Call "composer install" with a Docker image.

Add Docker image that builds the skin assets with node.js
Due to npm permission madness (npm needs to be run as root, but then all
its generated files are owned by root), we use a custom Dockerfile.

This allows the build machine ("local") to be free of application
dependencies (PHP with all its extensions, node.js). It only needs Git
and Docker.